refactor(fe): move validator logic to ValidatorPage after #3707 merge
